+++ .storybook.native/index.storybook.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import StorybookUIRoot from "./storybook"
+
+AppRegistry.registerComponent('storybooknative', () => StorybookUIRoot);