+++ packages/edition-slices/src/tiles/tile-ah/styles/index.js
@@ -29,7 +29,6 @@ const styles = {
     textAlign: "center"
   },
   imageContainer: {
-    borderRadius: 9999,
     overflow: "hidden",
     width: "40%"
   },