+++ packages/edition-slices/src/tiles/tile-an/index.js
@@ -16,6 +16,8 @@ const TileAN = ({ onPress, tile }) => (
       aspectRatio={1}
       style={styles.imageContainer}
       uri={getTileImageUri(tile, "crop11")}
+      rounded
+      resizeMode="cover"
     />
     <TileSummary
       headlineStyle={styles.headline}