+++ packages/edition-slices/src/tiles/tile-ap/index.js
@@ -15,6 +15,8 @@ const TileAP = ({ onPress, tile }) => (
       aspectRatio={1}
       style={styles.imageContainer}
       uri={getTileImageUri(tile, "crop11")}
+      rounded
+      resizeMode="cover"
     />
     <TileSummary
       headlineStyle={styles.headline}