+++ packages/jest-serializer/babel.config.js
@@ -1,0 +1,7 @@
+module.exports = api => {
+  api.cache(true);
+  return {
+    plugins: ["@babel/plugin-proposal-object-rest-spread"],
+    presets: ["module:metro-react-native-babel-preset"]
+  };
+};