+++ packages/message-bar/src/message-context.js
@@ -1,0 +1,6 @@
+import React from "react";
+
+export default React.createContext({
+  message: null,
+  showMessage: () => {}
+});