+++ packages/save-and-share-bar/src/save-and-share-bar.js
@@ -8,6 +8,7 @@ import {
   IconCopyLink,
   IconSaveBookmark
 } from "@times-components/icons";
+import withTrackEvents from "./tracking/with-track-events";
 import SharingApiUrls from "./constants";
 import styles from "./styles";
 import BarItem from "./bar-item";
@@ -45,8 +46,12 @@ class SaveAndShareBar extends Component {
   }
 
   render() {
-    const { articleUrl, onSaveToMyArticles } = this.props;
-
+    const {
+      articleUrl,
+      onSaveToMyArticles,
+      onShareOnFB,
+      onShareOnTwitter
+    } = this.props;
     return (
       <View style={styles.container}>
         <View style={styles.rowItem}>
@@ -59,6 +64,7 @@ class SaveAndShareBar extends Component {
             />
           </BarItem>
           <BarItem
+            onPress={onShareOnTwitter}
             target="_blank"
             url={`${SharingApiUrls.twitter}?text=${articleUrl}`}
           >
@@ -69,6 +75,7 @@ class SaveAndShareBar extends Component {
             />
           </BarItem>
           <BarItem
+            onPress={onShareOnFB}
             target="_blank"
             url={`${SharingApiUrls.facebook}?u=${articleUrl}`}
           >
@@ -109,11 +116,21 @@ class SaveAndShareBar extends Component {
   }
 }
 
+/* eslint-disable react/no-unused-prop-types */
 SaveAndShareBar.propTypes = {
+  articleId: PropTypes.string.isRequired,
   articleUrl: PropTypes.string.isRequired,
   articleHeadline: PropTypes.string.isRequired,
   onCopyLink: PropTypes.func.isRequired,
-  onSaveToMyArticles: PropTypes.func.isRequired
+  onSaveToMyArticles: PropTypes.func.isRequired,
+  onShareOnFB: PropTypes.func,
+  onShareOnTwitter: PropTypes.func
 };
 
-export default SaveAndShareBar;
+/* Serves as an indication when share links are clicked for tracking and analytics */
+SaveAndShareBar.defaultProps = {
+  onShareOnFB: () => {},
+  onShareOnTwitter: () => {}
+};
+
+export default withTrackEvents(SaveAndShareBar);