+++ ios-app/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.138](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.137...@times-components/ios-app@0.7.138) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.137](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.136...@times-components/ios-app@0.7.137) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.136](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.135...@times-components/ios-app@0.7.136) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.135](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.134...@times-components/ios-app@0.7.135) (2019-05-13)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.134](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.133...@times-components/ios-app@0.7.134) (2019-05-10)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.133](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.132...@times-components/ios-app@0.7.133) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.132](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.131...@times-components/ios-app@0.7.132) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 ## [0.7.131](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.130...@times-components/ios-app@0.7.131) (2019-05-08)
 
 **Note:** Version bump only for package @times-components/ios-app