+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.15](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.5.14...@times-components/article-comments@0.5.15) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
+## [0.5.14](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.5.13...@times-components/article-comments@0.5.14) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.5.13](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.5.12...@times-components/article-comments@0.5.13) (2019-05-08)
 
 **Note:** Version bump only for package @times-components/article-comments