+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.29](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.28...@times-components/article-extras@0.1.29) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.1.28](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.27...@times-components/article-extras@0.1.28) (2019-05-10)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.1.27](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.26...@times-components/article-extras@0.1.27) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.1.26](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.25...@times-components/article-extras@0.1.26) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.1.25](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.24...@times-components/article-extras@0.1.25) (2019-05-08)
 
 **Note:** Version bump only for package @times-components/article-extras