+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.2](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.2.1...@times-components/article-skeleton@1.2.2) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.2.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.2.0...@times-components/article-skeleton@1.2.1) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+# [1.2.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.1.0...@times-components/article-skeleton@1.2.0) (2019-05-13)
+
+
+### Features
+
+* add events tracking for save and share bar ([#1928](https://github.com/newsuk/times-components/issues/1928)) ([85348ca](https://github.com/newsuk/times-components/commit/85348ca))
+
+
+
+
+
+# [1.1.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.0.16...@times-components/article-skeleton@1.1.0) (2019-05-10)
+
+
+### Features
+
+* user notification bar (REPLAT-6144) ([#1896](https://github.com/newsuk/times-components/issues/1896)) ([90c0f49](https://github.com/newsuk/times-components/commit/90c0f49))
+
+
+
+
+
+## [1.0.16](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.0.15...@times-components/article-skeleton@1.0.16) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.0.15](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.0.14...@times-components/article-skeleton@1.0.15) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 ## [1.0.14](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.0.13...@times-components/article-skeleton@1.0.14) (2019-05-08)