+++ packages/lazy-load/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.31](https://github.com/newsuk/times-components/compare/@times-components/lazy-load@0.4.30...@times-components/lazy-load@0.4.31) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/lazy-load
+
+
+
+
+
+## [0.4.30](https://github.com/newsuk/times-components/compare/@times-components/lazy-load@0.4.29...@times-components/lazy-load@0.4.30) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/lazy-load
+
+
+
+
+
 ## [0.4.29](https://github.com/newsuk/times-components/compare/@times-components/lazy-load@0.4.28...@times-components/lazy-load@0.4.29) (2019-05-01)
 
 **Note:** Version bump only for package @times-components/lazy-load