+++ packages/save-and-share-bar/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.2.9...@times-components/save-and-share-bar@0.3.0) (2019-05-13)
+
+
+### Features
+
+* add events tracking for save and share bar ([#1928](https://github.com/newsuk/times-components/issues/1928)) ([85348ca](https://github.com/newsuk/times-components/commit/85348ca))
+
+
+
+
+
+## [0.2.9](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.2.8...@times-components/save-and-share-bar@0.2.9) (2019-05-10)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
+## [0.2.8](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.2.7...@times-components/save-and-share-bar@0.2.8) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
+## [0.2.7](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.2.6...@times-components/save-and-share-bar@0.2.7) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
 ## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.2.5...@times-components/save-and-share-bar@0.2.6) (2019-05-03)
 
 **Note:** Version bump only for package @times-components/save-and-share-bar