+++ packages/section/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.13](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.12...@times-components/section@0.13.13) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.12](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.11...@times-components/section@0.13.12) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.11](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.10...@times-components/section@0.13.11) (2019-05-10)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.10](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.9...@times-components/section@0.13.10) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.9](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.8...@times-components/section@0.13.9) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
 ## [0.13.8](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.7...@times-components/section@0.13.8) (2019-05-08)
 
 **Note:** Version bump only for package @times-components/section