+++ packages/video/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.21](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.20...@times-components/video@4.7.21) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
+## [4.7.20](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.19...@times-components/video@4.7.20) (2019-05-10)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
+## [4.7.19](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.18...@times-components/video@4.7.19) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
+## [4.7.18](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.17...@times-components/video@4.7.18) (2019-05-09)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 ## [4.7.17](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.16...@times-components/video@4.7.17) (2019-05-03)
 
 **Note:** Version bump only for package @times-components/video